requirements.txt
torch==1.12.1
torchvision==0.13.1
einops==0.3.0
timm==0.4.12
numpy==1.18.5
tqdm==4.54.0
wandb==0.10.23
matplotlib==3.3.3
scikit-image==0.18.1

Pillow>=8.1.1
h5py == 3.1.0
imageio == 2.9.0
opencv_python == 3.4.9.31
scipy == 1.5.2
scikit-learn

transformers == 4.16.2
utils == 1.0.1
Pygments>=2.7.4
timm==0.4.12
protobuf==3.20.3